fix(home): fix price
