+++ src/pages/home/Modules/Tabs/index.js
@@ -1,0 +1,2 @@
+import Tabs from "./container/TabsContainer";
+export default Tabs;
