+++ src/pages/home/Modules/TabsMB/index.js
@@ -1,0 +1,2 @@
+import TabsMB from "./container/TabsMBContainer"
+export default TabsMB
