+++ src/pages/home/Modules/Unit/container/UnitContainer.js
@@ -6,7 +6,23 @@ function UnitContainer() {
   const [data, setData] = useState([]);
   const getAllCourse = async () => {
     const { data } = await coursesApi.getCourses();
-    setData(data);
+    const newData = data.map((item) => {
+      return {
+        biDanh: item.biDanh,
+        danhMucKhoaHoc: item.danhMucKhoaHoc,
+        hinhAnh: item.hinhAnh,
+        luotXem: item.luotXem,
+        maKhoaHoc: item.maKhoaHoc,
+        maNhom: item.maNhom,
+        moTa: item.moTa,
+        ngayTao: item.ngayTao,
+        nguoiTao: item.nguoiTao,
+        soLuongHocVien: item.soLuongHocVien,
+        tenKhoaHoc: item.tenKhoaHoc,
+        price: 100,
+      };
+    });
+    setData(newData);
   };
   useEffect(() => {
     getAllCourse();
