+++ src/pages/home/Modules/index.js
@@ -1,5 +1,4 @@
 import React from "react";
-// import { getCourses } from "../../actions/courses";
 import BestCate from "./BestCate";
 import Billboard from "./Billboard";
 import ForBusiness from "./ForBusiness";
@@ -7,19 +6,17 @@ import HeadLine from "./HeadLine";
 import NonStudent from "./NonStudent";
 import OnBoarding from "./OnBoarding";
 import Partner from "./Partner";
-import TabCourseMobile from "./TabCourseMobile";
-import TabCourses from "./Tabs/components/TabCourses";
+import TabCourses from "./Tabs";
+import TabsMB from "./TabsMB";
 import TopicCate from "./TopicCate";
 import Unit from "./Unit";
-function Home({ courses }) {
+function Home() {
   return (
     <div className="home">
       <Billboard />
       <HeadLine />
       <TabCourses />
-      <div className="tabs-mobile">
-        <TabCourseMobile courses={courses} />
-      </div>
+      <TabsMB />
       <Unit />
       <OnBoarding />
       <BestCate />
